refactor(scoring): break the facade import cycle with a leaf fetchers module
_scores_raw.py no longer imports _logger from the package facade; it gets
its own logging.getLogger(__name__) (log records now carry the module's
own logger name instead of the facade's).

_max_history_runs and _make_trend_fetcher move out of __init__.py into a
new leaf module, _fetchers.py, which _project_scores.py imports and calls
through the module attribute (_fetchers._make_trend_fetcher(...)) rather
than binding the name locally. The facade still re-exports both names for
external importers and tests.

Both deferred, order-sensitive imports in __init__.py are now plain
top-of-file imports; no noqa: E402 left in the file.

Moved the services/scoring/__init__.py entry in
tests/tools/test_logging_boundary.py's DECLARED_LOGGING_SITES to
services/scoring/_scores_raw.py, since the logger moved with its only
consumer.

Fixed test_scoring_scalar_parity.py: patch.object(scoring,
"_make_trend_fetcher") was silently ineffective because _project_scores
bound the name at import time, so the "heavy path" branch of the test was
never exercised (both runs used the fast path). Now patches
_fetchers._make_trend_fetcher, which _project_scores actually reads at
call time.
