+++ tests/services/test_scoring_scalar_parity.py
@@ -26,10 +26,14 @@ def test_trend_identical_between_fast_and_heavy_paths(tmp_path: Path, monkeypatc
     fast = get_project_scores(reports, "proj")
 
     # Force the heavy path for the SAME data by making _make_trend_fetcher
-    # return the rescoring fetcher regardless of dismissals.
+    # return the rescoring fetcher regardless of dismissals. _project_scores
+    # calls through the _fetchers module attribute (not a name bound into its
+    # own namespace at import time), so patching it here actually takes
+    # effect -- patching the facade's re-export would not.
     import quodeq.services.scoring as scoring
+    from quodeq.services.scoring import _fetchers
     monkeypatch.setattr(
-        scoring, "_make_trend_fetcher",
+        _fetchers, "_make_trend_fetcher",
         lambda rr, p, params=scoring.DEFAULT_PARAMS, cacheable_run_ids=None, deps=None: (
             scoring._make_rescoring_fetcher(rr, p, params=params, deps=deps)
         ),
