+++ tests/tools/test_logging_boundary.py
@@ -105,8 +105,8 @@ DECLARED_LOGGING_SITES: dict[str, str] = {
     'services/mutation_rescore.py': 'Rescore-after-mutation helpers, shared by API routes and assistant actions - out of scope for this sweep (not a flagged per-site conversion)',
     'services/plugin_discovery.py': 'Discover available languages and return plugin metadata - out of scope for this sweep (not a flagged per-site conversion)',
     'services/score_run.py': 'Use case: score completed evidence after cancellation - imports the raw stdlib logger (out of scope for this sweep) and quodeq.shared.log_sink for quarantine-sink logging (log_malformed_jsonl_line/log_quarantined_findings) - plan-sanctioned composition wiring',
-    'services/scoring/__init__.py': 'Scoring reader - single read-side entry point for all score data - out of scope for this sweep (not a flagged per-site conversion)',
     'services/scoring/_rescoring.py': 'Accumulated-rescore machinery for the scoring reader - out of scope for this sweep (not a flagged per-site conversion)',
+    'services/scoring/_scores_raw.py': 'Single-run scores: SQL-backed with a JSON-eval-file fallback - own logger (quodeq.services.scoring._scores_raw), no longer threaded from the package facade (M-MOD-3 import-cycle fix) - out of scope for this sweep (not a flagged per-site conversion)',
     'services/shared_publish.py': 'Staging logic for publishing a project into the shared results repo - out of scope for this sweep (not a flagged per-site conversion)',
     'services/tooling_mixin.py': 'Mixin providing repo browsing and AI client discovery for the filesystem provider - out of scope for this sweep (not a flagged per-site conversion)',
     'services/violations.py': 'Violation resolution and aggregation for the filesystem action provider - out of scope for this sweep (not a flagged per-site conversion)',
